+++ R/plot_rich_norm.R
@@ -8,7 +8,9 @@
 #' @param facet.by A character string indicating whether the variable of
 #' interest is "density" or "threshold" (e.g. with DTI data)
 #' @param densities A numeric vector of the densities to plot
-#' @param alpha The significance level (DEFAULT: 0.05)
+#' @param alpha The significance level (default: 0.05)
+#' @param fdr A logical, indicating whether or not to use the FDR-adjusted
+#'   p-value for determining significance (default: TRUE)
 #' @param g A list (of lists) of \code{igraph} graph objects; required if you
 #'   want to plot a shaded region demarcating the \code{\link{rich.core}}
 #' @export
@@ -18,16 +20,20 @@
 #' @author Christopher G. Watson, \email{cgwatson@@bu.edu}
 #' @examples
 #' \dontrun{
-#' richPlot <- plot_rich_norm(rich.dt, densities[N:(N+1)], g=g)
+#' plot_rich_norm(rich.dt, facet.by='density', densities[N:(N+1)], g=g)
 #' }
 
 plot_rich_norm <- function(rich.dt, facet.by=c('density', 'threshold'),
-                           densities, alpha=0.05, g=NULL) {
+                           densities, alpha=0.05, fdr=TRUE, g=NULL) {
   p.fdr <- yloc <- Group <- phi <- xstart <- xend <- Study.ID <- NULL
 
   facet.by <- match.arg(facet.by)
   subDT <- rich.dt[eval(parse(text=facet.by)) %in% densities]
-  subDT[, star := ifelse(p.fdr < alpha, '*', '')]
+  if (isTRUE(fdr)) {
+    subDT[, star := ifelse(p.fdr < alpha, '*', '')]
+  } else {
+    subDT[, star := ifelse(p < alpha, '*', '')]
+  }
   subDT[, yloc := min(phi, na.rm=T) - 0.05 * diff(range(phi, na.rm=T)), by=facet.by]
   if (nlevels(subDT$Group) > 1) {
     for (i in 2:nlevels(subDT$Group)) {